+++ crates/jsonnet-statics/src/check.rs
@@ -159,11 +159,11 @@ pub(crate) fn get(st: &mut st::St<'_>, ar: &ExprArena, expr: Expr) -> ty::Ty {
     ExprData::Function { params, body } => {
       let mut param_tys = FxHashMap::<Id, ty::Ty>::default();
       // TODO type "annotations" via asserts
+      let m = def::ExprDefKindMulti::FnParam;
       for (idx, &(bind, rhs)) in params.iter().enumerate() {
-        let d = def::Def::Expr(expr, def::ExprDefKind::Multi(idx, def::ExprDefKindMulti::FnParam));
-        st.define(bind, ty::Ty::ANY, d);
+        st.define(bind, ty::Ty::ANY, def::Def::Expr(expr, def::ExprDefKind::Multi(idx, m)));
         if param_tys.insert(bind, ty::Ty::ANY).is_some() {
-          st.err(rhs.flatten().unwrap_or(expr), error::Kind::DuplicateBinding(bind));
+          st.err(rhs.flatten().unwrap_or(expr), error::Kind::DuplicateBinding(bind, idx, m));
         }
       }
       for &(_, rhs) in params {
@@ -292,10 +292,10 @@ fn define_binds(
   m: def::ExprDefKindMulti,
 ) {
   let mut bound_ids = FxHashSet::<Id>::default();
-  for (idx, &(bind, rhs)) in binds.iter().enumerate() {
+  for (idx, &(bind, _)) in binds.iter().enumerate() {
     st.define(bind, ty::Ty::ANY, def::Def::Expr(expr, def::ExprDefKind::Multi(idx, m)));
     if !bound_ids.insert(bind) {
-      st.err(rhs.unwrap_or(expr), error::Kind::DuplicateBinding(bind));
+      st.err(expr, error::Kind::DuplicateBinding(bind, idx, m));
     }
   }
   for &(lhs, rhs) in binds {