+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix types for `std.slice` params.
 - Improve hover for `local` binds.
+- Improve go-to-def, etc. locations for `local f(x) = ...` function parameters.
 
 ## v0.5.0